src/detector/requirements.txt
absl-py==0.9.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
google-auth==1.10.0
google-auth-oauthlib==0.4.1
grpcio==1.26.0
idna==2.8
joblib==0.14.1
kafka-python==1.4.7
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.2
numpy==1.18.1
oauthlib==3.1.0
pandas==0.25.3
Pillow==7.0.0
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22.1
scipy==1.4.1
six==1.13.0
tensorboard==2.1.0
torch==1.3.1
torchvision==0.4.2
urllib3==1.25.7
Werkzeug==0.16.0
pyspark==2.4.4
